requirements.txt
flask
Flask-SQLAlchemy
flask-admin
flask-security
Flask-Script
sqlalchemy
passlib
bcrypt
pandas
python-dateutil
xlrd
tqdm
natsort